исправил pr2d_Circle и pr2d_Ellipse; исправил проблемы отсечения невидимых спрайтов при CORRECT_RESOLUTION; text_DrawInRect теперь учитывает символ переноса строки LF(10)

+++ src/zgl_opengl_simple.pas
@@ -45,8 +45,8 @@ uses
   zgl_opengl_all;
 
 var
-  tSCount  : WORD;
-  tScissor : array of array[ 0..3 ] of WORD;
+  tSCount  : Integer;
+  tScissor : array of array[ 0..3 ] of Integer;
 
 procedure Set2DMode;
 begin
